set up folder structure, add login and sign up pages
