+++ src/pages/index.tsx
@@ -1,0 +1,4 @@
+export * from './homePage';
+export * from './errorPage';
+export * from './loginPage';
+export * from './signupPage';
