+++ src/pages/loginPage/LoginPage.tsx
@@ -1,0 +1,3 @@
+export const LoginPage = () => {
+    return <div>This is Login Page</div>;
+};
