+++ src/pages/signupPage/SignUpPage.tsx
@@ -1,0 +1,3 @@
+export const SignUpPage = () => {
+    return <div>This is SignUp Page</div>;
+};
